fix: fixed all the generic value comparisons in the code.
Since the generic object types can't be compared using relational operator
as done in the code before this commit. Hence all the comparisons have
been done using the compareTo() method. For this method to be used on the
generic object type, the object type is made to extend the INTERFACE
Comparable.